Remove unnecessary route
We had a failing spec on CI few days ago saying that it couldn't
find sign_out route. We added this temporary to test whether
it was because it couldn't find the route. Finally we found that
this error was due to the fact we had the Log out as link_to
method and not button_to