Add a drawing numbering step to assessments
This new step for assessors prepares the application for
submission to their manager, and ultimately, publication. Drawing
numbers are displayed in the decision notice page and email.

It should not be possible to submit the assessment until all drawings
for publication have numbers, but that is not yet enforced at the
model / controller layer.

Co-authored-by: Rhian Lewis <rhianatwork@gmail.com>