Raise error if problem in downloading plans via API
We were silently swallowing errors if a plan couldn't be downloaded that was sent to us via the API. We want it to explode and let everyone know if this happens.

In the future, we should add some retrying and handle it more gracefully, but this allows us to quickly identify problems during this early period.

I have moved the HTTP stubbing to only the tests that it is relevant to, this makes it easier to overwrite for checking errors, as well as making the tests more explicit.

I have also made the stubbed successful request return a valid file rather than an empty string. This was partly because we now explode if it's not a valid file, but also to explicitly test that the file is loaded correctly. This slows down the tests a bit, but I think makes them more resilient. If we want to speed this up in the future, we can find a new test file that is smaller.