Move swagger_doc.json into public folder
swagger_doc.json wasn't being served on preview. I've tried looking into how rswag hooks into the routes to server it, and why it works locally, and not on production, but I can't figure it out, and I need this to go live today.

So I've moved the json file into the public folder, and updated the path for the rswag configuration.