Stop using rswag to generate oas3 document
Using rswag to generate the swagger_doc was getting quite cumbersome, so I have switched tacts: I am manually editing the swagger doc, and have added some specs to validate the swagger doc is accurate.

I have done this by using openapi3_parser gem to parse the document, grab out the examples examples in the document, and then post that example to the test server during the spec run. This ensures that the document is valid, and also our stated examples really do work as expected.

I have also removed rswag-specs as we're not using it any more, and rswag-api as we can keep the document in public. I have left rswag-ui for now, but we might find a lighter weight solution in the future.