Add document validation step
This adds a form for the assessor to validate the documents. The application cannot transition to in_assessment without the documents_validated_at field being populated. The assessor also has the option to invalidate the application, and in each case the planning application appears correctly flagged in the index table