Rename Drawings to Documents everywhere
We want to be more generic when talking about documents. Documents could be drawings, but they also could be invoices or other kinds of documents. At the moment, we don't have a way of distinguishing between these different types of documents, which we may do in the future, but for now, we will call them all documents.