Update specs for UI permissions in pre-assessment statuses
Planning officers should be able to see Upload Documents button if application is in any status other than Closed. They should have read-only access to the assess documents form unless application is in in_assessment status. These specs verify this, as well as the validation message that is displayed when date is missing from validation form