Update target_date when documents validated
Before validation, the target date is 8 week from the creation of the application, but this might change when the documents are validated. This updates the target_date.

Implimentation note: I am using both `before_create` and `before_update` here, as `before_save`, in the current version of Rails, happens before the created_at timestamp is created, which results in trying to do maths on nil.