Set extract_css to true in development
This is in prep for PDF rendering.

By default, webpacker loads all of the CSS via a javascript loader in development mode. I believe this is to allow faster rendering times when you make a change to the CSS. However, loading CSS through javascript makes me anxious, and was indeed causing problems for PDF rendering in development as the CSS wasn't being loaded.