Exclude PDFKit::Middleware from active_storage paths
The PDFKit::Middleware was doing... something on the active_storage URLs for uploaded PDF files. It wasn't actually erroring anything, but was adding 5% more load time for these URLs.

This was causing our documents system tests to fail because they were going quicker than the application. This is a bad sign for these system tests that they are that flakey, but in this particular instance, it helped catch that we probably should exclude active_storage from the PDFKit Middelware.

So to summerize:

* I don't know what PDFKit::Middleware was taking so long about with active_storage PDFs, but it shouldn't be doing it anyway, so disable
* This brings our flakey documents system tests to pass, but we should make these more resillient anyway.