Remove production default url options for mailer
In 0879918dc3da9891c6b239a03897e7ea75c06ea1 I ensured that we passed the request host through to any mailers we sent, given that we are multi-tenant, and therefore we don't know the host from the environment.

However, I missed that this default_url_options had been set in production.rb, so the host was being set correctly, but the port was being set as 3000 by default, which is wrong.

I believe that removing it is the best solution: I would rather error than have an incorrect URL silently get sent.