Only show documents that are for publication in the API
Previously we were revealing all documents through the API. We only want to show documents that are "for publication" (i.e. have been given a number and have not been archived). We are removing the archived_at and archive_reason from the API too, as they will never not be nil now.