Redesign/refactor task list
The task list was confusing, both to users and the codebase. This is a complete rewrite and redesign of the task list.

There are now 5 static steps that are consistent. What changes is whether they are accessible, or whether they are marked as completed. The logic for this is moved into the planning application model. I have created a feature spec for the task list which specififes every state the task list should be in for every state we currently have.