Remove pundit
At this point, pundit is doing nothing, except scoping planning applications by local authority on the index page.

We are yet to add back in the restrictions, but there is only one: assessors cannot review applications. This can be done with a simple controller check, so using pundit is overkill.

In addition, the extra code overhead that pundit was bringing meant that an oversight was made: it was still possible to visit any action on a planning application that isn't the index page!

* remove pundit
* scope all calls for finding a planning application by the current local authority
* fix a lot of system specs that were correctly failing, as they were accessing planning applications that were in a different local authority
* add a couple of quick request specs on planning application show page and documents index page
* disable some false positives that brakeman was raising
* re-organise request specs to scope api request specs more obviously