Remove link from reviewer actions when assessor is viewing
We should not show the links for Review and Publish to an assessor so they cannot acces the view. For the moment, we are not restricting access to the page itself as it is an internal system. Specs for what the assessor sees on the task list have been added to task_list_spec.rb