Replace create_sample_data with simpler seeds file
The create_sample_data rake task would not work with the new recommendations model. However, now that we have the API with example data in the swagger documentations, we don't need it to generate example data.

Therefore, rather than update it, I have removed all the example planning applications, leaving only the creation of models that are required to get the application running, which means it can be a seed file again.