Collapse statuses i18n hash
There was a lot of i18n information that was being used as a way to change the content on the front end based on the planning application. This isn't really what i18n is for.

We have now mostly removed all of this, so we can now remove it from en.yml, leaving it as just a simple translation between how we store the status in the database, and how we want users to read it in english.