Display all the metadata in the proposal_details on show page
For now, I am simply and basically exposing everything we might receive from RIPA. This might not be how we want to do it, based on what the data we start receiving might look like, but allows users to see that data.

Given I don't have a spec of what this *should* look like, I have not written specs for it yet. Once the feature has settled down a bit, this should be done.