Fix deprecation warnings
- Errors no longer accept two arguments, this commit fixes this deprecation warning:
You are passing a block expecting two parameters, so the old hash behavior is simulated. As this is deprecated,
this will result in an ArgumentError in Rails 6.2. (called from views/planning_applications/recommendation_form.html.erb:16)
.DEPRECATION WARNING: Enumerating ActiveModel::Errors as a hash has been deprecated.
- Rendering actions with '.' in the name is deprecated so all partials with .html or .json now have the simplified naming convention
- raise_on_missing_translations has been replaced with i18n_raise_on_missing_translations due to deprecation
- host! has been deprecated with no replacement - the only place where we don't get any deprecation warnings is in the "request" setting in rails helper, where we do need it or all api requests fail.
I could not find an alternative to host! that would force the GET URI to not have www and since no warnings appear it seemed ok to keep.