Add boundary map to accordion
As the accordion starts off closed on page load, the leaflet map will load incorrectly due it's size being 0. `invalidateSize()` needs to be called on the map once it is visible on the page.

This has required me to monkey patch the GOV.UK frontend accordion to give me an event function that can be called to do this. I'm not sure this is the best way to achieve either this monkey patch or the event function, but it works.