Ensure map doesn't overlap page break
Sometimes in the PDF rendering (which uses print CSS), the map would overlap a page break, causing the map to be split over 2 pages.

I've added `page-break-inside: avoid;` to the map which stops this happening! (new to me).

I have also put a wrapping container on the map and added padding to that, otherwise the map hugs the very top of the page.