planning application: introduce policy references
Planning applications need to conform to the General Permitted
Development[1], which is ordered in: schedules (Schedule 2) containing
parts (Part 1) containing classes (Class A) containing
references/paragraphs[2].

To ease the process of checking applications against it, we want to
let officers add some parts/classes to an applicaton and be able to
track whether the specific points/references within that class comply,
don't comply or are yet to be determined.

To avoid modeling the whole tree of Schedule -> Part -> Class ->
References as ActiveRecord tables that we would need to quadri-join,
and maintain as the legal text changes (about twice a year), we use a
big YAML file to store and grab all of this tedious copy/hierarchy,
and stamp it as JSONB in the PlanningApplication when needed.

This means that in the case of the legal text being updated in the
app, all previous applications will still hold the "old" copy (since
all of it goes in that JSONB column), avoiding a dangerous drift of
content between the new version and the applications in progress.

[1]: https://www.legislation.gov.uk/uksi/2015/596/contents/made
[2]: https://www.legislation.gov.uk/uksi/2015/596/schedule/2/made