policies: introduce a wrapping model for policy classes
Register a custom type to easily cast the policy classes into their
own PolicyClass wrapper model. This makes it easier to add some of the
fiddly logic like:

* stamping the part within the policy class: because our YAML policies
are organised in a logical tree of schedules -> parts -> classes ->
policies, when we cherry pick which classes we want to assess against
we need to remember which part it was linked to;
* stamping the policies status: once a policy class is registered, all
its policies need a default status of "undetermined".

Note that because of a bug, or me misunderstanding how Rails seralises
JSONB arrays, this line in the policy_assessments_controller.rb is
crucial:

@planning_application.policy_classes_will_change!

Without it the policies statuses, despite being evidently changed in
memory, are not recorded/persisted as part of
`@planning_application.save`.