policy class: ensure policy references are cast as strings
One of our policies looks like this (class E):

```
-
  id: 3
  description: |
  In the case of any land within the curtilage of [...]
```

So far so good but YAML will parse one or more digits as
numbers, which in hindsight makes sense.

```
irb(main):026:0> YAML.load('id: 3')
=> {"id"=>3}

irb(main):027:0> YAML.load('id: 3b')
=> {"id"=>"3b"}
```

This means that when a reviewer is trying to assess the class
policies and Rails send string params over:

```
{"policies"=>{"1a"=>"complies", "3"=>"complies"}, ... }
```

there's a mismatch in trying to find the corresponding values in the
params. Work around it by always casting the policy ID to a string
when digging into the params.