e2e: add some end to end tests
It's all a bit sketchy at the moment but in a nutshell:

When we run E2E tests we want Capybara to boot a server as always, but
then it needs to go to another app (applicants) that redirects to
our *test* BOPS (i.e the one that's running on the Capybara
server). If we redirect to 3000, that's our dev server so the data
doesn't match.

So pin our E2E tests server (only fired when tagged with '@e2e', the
default driver rack-test does not boot a server, at all) to port 3333
and reflect that in the docker-compose config to line up a test
pipeline that can successfully communicate.

It needs cleaning up but it's working (!)