controllers: fix flash usage
If we set the flash and then use `render :some_template` then flash is
set for the request and the next one - that's what the flash does. So
if we re-render a create form with `flash[:error]` set, if the form is
submitted successfully then the next request still has the flash.

Fix this by using flash.now instead, and also use the shortcut syntax
to clean up the code a bit.