planning application: add a decorator
We have way too much logic into the model which we've been trying to
extract, initially with a presenter. The issue with the presenter is
that it's tied to a view context (it's initialised with one) so that
it can wield some slightly more complicated HTML tags:

tag.strong "foo" do
  tag.span "bar"
end

This is great and makes the view a lot cleaner but it's not the right
place to store methods that just want to format the model's attributes
and return scalar values (i.e simple types: strings, numbers,
booleans - not HTML tags or other compound things), especially in
places where you might not have access to a view context (an audit
entry for example).

For these methods introduce a decorator that's included straight away
into the application which means the methods are immediately
available, without having to construct and wrap a presenter around the
class.

This decorator can hold methods like:

def full_address
  "#{address_1}, #{town}, #{postcode}"
end

which is just a string and doesn't need to be aware of a view
context.

This pattern can be abused if we start stuffing the decorator with
random bits and bobs (which is why concerns get a bad rep in general)
but here's the rule of thumb:

- model: pure/direct logic (state, updates, etc);
- decorator: simple, scalar computations;
- presenter: markup logic.

Eventually we might revisit this approach and switch to view
components, or decorators that aren't mixed in straight away, i.e you
have to call application#decorate to get the good stuff, but this
should carry us to this place very nicely.