pdf: swap pdfkit for grover
We want to include the planning application's digital map (if present)
in the PDF we generate for the decision notice.

We were doing so using pdfkit[1] which relies on wkhtmltopdf[2] which
is basically a wrapper around QtWebkit (a headless browser) except it
uses an old version of QtWebkit[3] that does not support modern ES6
features. Why do we need modern ES6 features? Because we use OSL's
map[4] library and they define a custom compoment and that's all too
modern for QtWebkit which will choke on it (visible if wkhtmltopdf is
passed the `--log-level info` switch).

Their recommendation page actually says:

> If you’re using it to convert a site which uses dynamic JS, consider
using puppeteer or one of the many wrappers it has.

Instead, turn to Grover[5] which is a gem that leverages puppeteer to
generate the PDF, and provides a similar middleware to the one that
PdfKit offered.

PS: this will need the relevant changes in bops-terraform to provide
the tooling in production too.

[1]: https://github.com/pdfkit/pdfkit
[2]: https://github.com/wkhtmltopdf/wkhtmltopdf
[3]: https://wkhtmltopdf.org/status.html#recommendations
[4]: https://github.com/theopensystemslab/map/
[5]: https://github.com/Studiosity/grover