constraints: add a note about the no-mapping against the API
We don't seem to be doing any form of lookup on the values that come
from the API - there's this method:

      def constraints_array_from_param(constraints_params)
        if constraints_params.present?
          constraints_params.to_unsafe_hash.collect do |key, value|
            key if value
          end.compact
        else
          []
        end
      end

but it doesn't seem to do anything besides removing empty values.

We do store them, but the RIPA constraints and our constraints did not
and still don't match. For example, they're missing a bunch of
ours (Military, ASNW, Green Belt, etc etc), and we're missing a bunch
of council-specific ones (i.e: article4.lambeth.stockwell).

Ideally we'd align on that, but I guess the constraints aren't used so
much at this point. When we decide to align, it's just a matter of
adding the correct mapping in the YAML file, so that instead of:

    heritage_and_conservation:
      - Listed Building

it is

    heritage_and_conservation:
      listed: Listed Building

with "listed" being the constraint name sent over by RIPA.

Anyway, whoever reads this, good luck!