Update to Rails 7.0.4.2 and upgrade dependencies
- Update Rails JavaScript packages
- Run bin/rails app:update

Due to a breaking change for 2FA this also introduces a bump in our devise-two-factor gem version. The older version will not work on Rails 7 and upgrading to 5.x for this gem requires Rails 7 as a dependency... So it needs to be done together as one upgrade

Steps to upgrade are followed here: https://github.com/tinfoil/devise-two-factor/blob/main/UPGRADING.md#upgrading-from-4x-to-5x-1

Versions 5+ of this gem uses a single Rails 7+ encrypted attribute named otp_secret to store the OTP secret in the database table so this adds environment specific rails credentials.

Details in the README are added to get 2FA working in development

We will need a future migration to clean up as a separate deployment but this is outlined in the upgrading docs above