Fix warning that img/bmp is not a valid content type
DEPRECATION WARNING: image/bmp is not a valid content type, it should not be used when creating a blob, and support for it will be removed in Rails 7.1. If you want to keep supporting this content type past Rails 7.1, add it to `config.active_storage.variable_content_types`.

Since we are just testing that an invalid file cannot be uploaded, this change introduces the .gif extenstion to test with