Rename constraints.yml to avoid mixup
"Constraints" are defined in two places in the various yml files,
resulting in when we iterate through the list of constraints for the
constraint edit page we get some weird results - the success/update
messages are listed as actual constraints.

Rename the constraint.yml file so this doesn't happen. Thought that
was easier than trying to override the constraints controller to know
to look for something different.