Fix assessment against legislation intermittent spec
The intermittent failure was due to a Selenium::WebDriver::Error::UnexpectedAlertOpenError caused by navigating away from the page due to the unsaved_changes_controller where it checks if you will leave the page with any unsaved content. Occasionally the current_path remains on the policy classes page and therefore it cannot find the link "Part 1, Class D". This might actually be a bug with Selenium as the failed screenshot suggests this shouldn't be the case.

This fix just splits the test where the important thing to check is that the updated comment and previous comments are visible