Create database structure for constraints
- Adds Constraints table
- Adds PlanningApplicationConstraintsQuery table
- Adds PlanningApplicationConstraint table

We used to have constraints as an array field on planning applications. Thus, as we mentioned on the comments bellow we agreed on migrating the constraints array to "old_constraints", so we don't need to do any further migrations after this PR.

However, since we will be making an API request to retrieve the updated constraints from planx/planning data, we need to modify the schema to allow a relation with the query made and constraints returned.

This commit also migrates the custom/local constraints of the existing "old_constraints" array on planning applications to the new association. This has been saved as "constraints".

The constraints_list.yml has been kept in for now.

Since custom/local constraints can be added, we need to iterate and migrate the existing items in all the planning application constraints array that include a name that hasn't been seeded and create a new Constraint entry in the database. This needs to be done before we can switch the code to use the new has_many constraints association for planning applications.

In addition we will add an optional reference of local_authorities to constraints as they might have their own constraints, thus we will add the enum local.