Add infrastructure for controlling reply-to addresses
GOV.UK Notify reply-to addresses are specified using UUIDs created by the
system so we need to pass those in the email payload. Default to a per-LA
model for now but add the columns so that we can add multiple options for
consultation reply-to addresses at some point in the future.