Launch foreman using the inbuilt tini init process
Whilst it appears that the grover gem is using Open3.popen3 correctly
the puppeteer npm package it uses internally is somehow sending TERM
signals that kill the foreman process. By wrapping it with the tini
init daemon that is inbuilt to Docker since 1.13 we can prevent these
child/zombie processes from killing foreman whilst at the same time
ensure that ctrl-c is forewarded correctly so that the containers
will stop correctly.