Make consultation start/end dates consistent
The start date time should be set to the beginning of the day and the
end date time should be set to the end of the day. Also floor the days
left calculation so it always changes at midnight and not at noon.

Note that on the last day of the consultation it will show zero days
remaining. If that's not required all that need to change is the floor
should be changed to ceil.