Switch away from native interpolation and validate fields
The built-in Ruby `format` method is inappropriate for use on user-supplied data
as it can use a variety of formats and generate exceptions when substitutions
are improperly formed. Use `gsub` instead which will ignore anything that doesn't
match the pattern and validate the message fields on submission for any placeholders
that don't match expected values.