Allow any subdomain to be used
Makes testing onboarding easier and allows for additional engines
on non-LPA subdomains such as admin.bops.localhost.

Conventionally a subdomain of localhost will resolve to 127.0.0.1
for IPv4 and ::1 for IPv6. This works fine for browsers outside of
the docker network but where we have containers talking to each
other such as bops-applicants querying the API of bops this will
fail since 127.0.0.1 will point to the originating container.

To work around this limitation we need to override the default
behaviour to make DNS requests for bops and bops-applicants return
their respective container IP addresses. For a limited number of
hosts, using network aliases works but becomes cumbersome quickly
and network aliases don't support wildcards.

The solution to this is to add our own DNS service to the network
and dnsmasq is ideally suited to this task. Since we need to specify
the IP addresses in the config we need to assign static addresses
to the bops and bops-applicants services (and also dnsmasq itself).

Now that bops and bops-applicants have static IP addresses we need
to add additional services for running tasks since we can't run
multiple instances of the existing services. To facilitate this we
extract out common config to reusable YAML blocks. We also need to
make sure that the new 'console' services don't boot when we run
'docker compose up', so we use the docker compose v2 profile feature
to disable them by default and use '--profile console' when running
tasks such as 'make lint' or 'make migrate'.