Allow validation when description change exists
Fun bug. If there were any open validation requests then the system
thought we needed to invalidate the application. However, when you
did that, because the description change wasn't in a pending state, it
said you couldn't invalidate it becasue there weren't any pending
validation requests.

This fixes it so that you can validate with an open description change
request.