Ensure application creation date is before review date
Because the application was created outside the `travel_to`, it was
getting a creation date based on the current date as the test was run.
The review, however, happened at a specified date in the future: but
once that date was passed, the review would again be happening before
the application had been created and the test would fail. This led to
1e241ba7 where the date was updated to be in the future once more.