Don't raise an error when max attempts have been exceeded
The default for `retry_on` is to retry for five times and then raise an
error but this triggers a doom loop with the default of retrying a job
when it fails with an unhandled exception.

Fix this by setting the number of attempts to twelve and then using a
block to send an error to Appsignal. This gives a job an hour to be
successfully finished - if it's not then it's likely something is down
somewhere.