Restart web and applicants if they fail
Sometimes one of the background processes like yarn causes a service to
exit and you have to manually stop the stack and restart it. This seems
to be mainly on first booting the stack after docker has started and
is visible in the log output like this:

```
An unexpected error occurred: "ENOSYS: function not implemented, fstat".
```

This appears to be down to the filesystem not being fully initialised
in the container so it's fine to just restart the container.