Make informatives match the Figma design
Wrap the add new informative form in a details element that is open by
default when there are no informatives or there are errors, but closed
by default when there are informatives present.

Also adds a section break between the add informative form and the
button group at the foot of the page.