Be more specific in application type `consultation?` and `consultation_steps`
If the steps contain 'consultation' but there are no consultation steps
defined, this method will return false. This means that a consultation
record will not be created for an application of this type. However,
when iterating over the steps, it will include 'consultation' and thus
try to render those tasks, which will fail.

So, these should behave consistently, it should not be possible to have
application types with consultations but no consultation steps or vice
versa.