Show a link to document validation request where appropriate
Currently we're always showing a link to request a replacement even if
one is already in the process of being requested. This also makes it
hard to cancel requests after the validation stage (when the
validation request view isn't so easily accessible). Linking to the
management page for an existing request when one exists seems reasonable.