Replace Action Text JS with our own
The url /blobs/redirect/:signed_id/*filename never hits the app on staging / production as
this is reserved for accessing blobs from S3 directly. To work around this replace the
Action Text JS lib with our own derived from it the source but replacing the function that
creates the blob url with our own based on the key. This means we can remove the redirect
route and the associated controller.