Update accessibility statement
Primarily this has just meant updating the dates, but on reviewing the
last accessibility report I've also downgraded the claim from AA to A,
because we have a few outstanding AA tasks that we never completed. A
couple of these relate to the map, which we have also declared is out
of scope, but there's also an issue with the unclear default error messages.