Remove dead link
In theory we could update this to the latest version of the report but
this was already dead and we'd need to figure out a new place to serve
it from (that domain now redirects to ODP). Suggest we revisit when a
new report is done / if anyone requests it.