Add additional expectations around non-delivery of messages
There was a report that a report email was sent after submitting a recommendation.
It was not possible to replicate this so extra expectations were added to the self
assess and review flow to ensure it's only sent after being determined.

Also removes the publish action added in #2350 as it's no longer required.