Reimplement GOV.UK Notify checks as a multi-step form
This saves any changes to the GOV.UK Notify settings first and then goes
through email, sms and letter forms to send or create test messages/letters.

Note that this create and enqueues a real letter - this needs to be cancelled
by someone with GOV.UK Notify access if this is not to incur charges. Whilst
it would be preferable not to do this the template preview function doesn't
provide enough feedback to ensure the letter has been created successfully.