Fix double application of strong_params filtering on task params
It's called once in the controller and once in the form model, the
second time on params that have already been filtered. Limiting it to
the controller, because that seems to be the more common pattern than
leaving them unfiltered in the controller and putting the filter in
the form model instead.