Add additional breakpoint for when sidebar is present
The content was losing the margin when on the last breakpoint due to it
being set to 'auto'. Fixed by overriding it to restore the 30px margin
and adding an extra breakpoint at the old width + sidebar width.

Also move the CSS variable to a Sass variable due to scoping issues.