Enable new sidebar for assessment in production
Each controller/action that needs to show the sidebar can call the
`use_new_sidebar_layout` method with a symbol (:validation,
:assessment, etc), and it will return a boolean.

The configuration parameter can now be set to true or false as before,
but also to a symbol or array of symbols representing the stages of
the application for which it should be enabled.

Currently set to `:assessment` in production and `true` in staging.